+++ libraries/AP_BLHeli/AP_BLHeli.h
@@ -53,7 +53,7 @@ public:
         return channel_bidir_dshot_mask.get() & (1U << motor_map[esc_index]);
     }
 
-    uint16_t get_bidir_dshot_mask() const { return channel_bidir_dshot_mask.get(); }
+    uint32_t get_bidir_dshot_mask() const { return channel_bidir_dshot_mask.get(); }
 
     static AP_BLHeli *get_singleton(void) {
         return _singleton;
@@ -232,7 +232,7 @@ private:
     // have we disabled motor outputs?
     bool motors_disabled;
     // mask of channels that should normally be disabled
-    uint16_t motors_disabled_mask;
+    uint32_t motors_disabled_mask;
 
     // have we locked the UART?
     bool uart_locked;
